requirement.txt
appdirs==1.4.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
cssselect==1.1.0
dateparser==1.0.0
demjson==2.2.4
dnspython==1.16.0
facebook-scraper==0.2.44
fake-useragent==0.1.11
idna==2.10
lxml==4.6.3
numpy==1.21.0
pandas==1.3.0
parse==1.19.0
pyee==8.1.0
pymongo==3.11.4
pyppeteer==0.2.5
pyquery==1.4.3
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
regex==2021.7.6
requests==2.25.1
requests-html==0.10.0
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
tqdm==4.61.2
tzlocal==2.1
urllib3==1.26.6
w3lib==1.22.0
websockets==8.1
